+++ ml_project/tests/features/test_outlier_transformer.py
@@ -1,55 +1,10 @@
-import pytest
 import pandas as pd
 
-from src.entities.feature_params import FeatureParams
-from typing import List
-
 from src.features.outlier_transformer import OutlierTransformer
+from src.entities import FeatureParams
 
 
-@pytest.fixture()
-def test_df() -> pd.DataFrame:
-    df = pd.DataFrame({
-        "feature": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
-    })
-    return df
-
-
-@pytest.fixture()
-def categorical_features() -> List[str]:
-    return ['']
-
-
-@pytest.fixture()
-def numerical_features() -> List[str]:
-    return ['feature']
-
-
-@pytest.fixture()
-def features_to_drop() -> List[str]:
-    return ['']
-
-
-@pytest.fixture()
-def target_col() -> str:
-    return ''
-
-
-@pytest.fixture()
-def feature_params(categorical_features: List[str],
-                   numerical_features: List[str],
-                   features_to_drop: List[str],
-                   target_col: str,
-                   ) -> FeatureParams:
-    return FeatureParams(
-        categorical_features=categorical_features,
-        numerical_features=numerical_features,
-        features_to_drop=features_to_drop,
-        target_col=target_col,
-    )
-
-
-def test_outlier_transformer(test_df: pd.DataFrame):
+def test_outlier_transformer(test_df: pd.DataFrame, feature_params: FeatureParams):
     transformer = OutlierTransformer(feature_params)
-    transformed_df = transformer.transform(test_df, columns=['feature'])
+    transformed_df = transformer.transform(test_df)
     assert transformed_df.shape == test_df.shape
